refactor(editor): rename styles.ts to module.style.ts
