+++ packages/brickdoc-editor/src/components/blockViews/EmbedView/Placeholder.tsx
@@ -1,7 +1,7 @@
 import { FC, ReactNode } from 'react'
 import { styled, theme } from '@brickdoc/design-system'
 import { UploadProgress } from '@brickdoc/uploader'
-import { maxWidth } from './styles'
+import { maxWidth } from './EmbedView.style'
 
 export interface EmbedBlockPlaceholderProps {
   icon: ReactNode
