+++ packages/brickdoc-editor/src/components/blockViews/EmbedView/embedTypes/Gallery/Gallery.style.ts
