+++ packages/brickdoc-editor/src/components/blockViews/EmbedView/embedViews/CardView/CardView.style.ts
@@ -1,5 +1,5 @@
 import { styled, theme, Button } from '@brickdoc/design-system'
-import { maxWidth } from '../../styles'
+import { maxWidth } from '../../EmbedView.style'
 import { EmbedToolbar } from '../EmbedToolbar'
 
 export const EmbedToolbarContainer = styled('div', {
