+++ packages/brickdoc-editor/src/components/blockViews/EmbedView/embedViews/PreviewView/PreviewView.tsx
@@ -8,7 +8,7 @@ import { PdftronDocument } from './PdftronDocument/PdftronDocument'
 import { WebsiteDocument } from './WebsiteDocument'
 import { styled } from '@brickdoc/design-system'
 import { EmbedToolbarContainer } from './DocumentFooter'
-import { maxWidth } from '../../styles'
+import { maxWidth } from '../../EmbedView.style'
 
 export interface PreviewViewProps {
   blockType: EmbedBlockType
