+++ packages/brickdoc-editor/src/components/blockViews/ListView/ListView.tsx
@@ -3,7 +3,7 @@ import { NodeViewContent } from '@tiptap/react'
 import { BlockContainer } from '../BlockContainer'
 import { BlockViewProps } from '../../../extensions/common'
 import { BulletList, OrderedList, TaskList } from '../../../extensions'
-import { orderedListStyles, bulletListStyles, taskListStyles } from './styles'
+import { orderedListStyles, bulletListStyles, taskListStyles } from './ListView.style'
 
 export const ListView: FC<BlockViewProps<{}, {}>> = ({ deleteNode, node, getPos }) => {
   const as = node.type.name === OrderedList.name ? 'ol' : 'ul'
