+++ packages/brickdoc-editor/src/components/editorViews/ExplorerMenu/ExplorerMenu.tsx
@@ -11,7 +11,7 @@ import {
   MenuItem,
   SearchInput,
   SearchInputContainer
-} from './styles'
+} from './ExplorerMenu.style'
 import { Drawer } from '../../ui'
 import { useDrawer } from '../../ui/Drawer'
 import { slashMenuGroup } from '../../../extensions/extensions/slashCommands/items'
