+++ packages/brickdoc-editor/src/editors/commentEditor/commentEditor.style.ts
