+++ packages/brickdoc-editor/src/editors/commentEditor/commentEditor.tsx
@@ -3,7 +3,7 @@ import { Editor, JSONContent } from '@tiptap/core'
 import { Button } from '@brickdoc/design-system'
 import { BrickdocEventBus, DiscussionMarkInactive } from '@brickdoc/schema'
 import { usePlaceholder } from './usePlaceholder'
-import { EditorContainer, EditorAvatar, EditorInput, ActionsRow } from './styles'
+import { EditorContainer, EditorAvatar, EditorInput, ActionsRow } from './commentEditor.style'
 import { useCommentEditor } from './useCommentEditor'
 import { clearDraft, getDraft } from './draft'
 import { useContentUpdated } from './useContentUpdated'
