+++ packages/brickdoc-editor/src/editors/documentEditor/documentEditor.style.ts
